chore(HIGH-56): 쇼츠 나가기 버튼 위치 수정
